refactor(cli): extract manual.py DB error classifier to _manual_errors
Moves the two DB error helpers (_summarize_db_exception and
_report_manual_db_error) out of amx/cli_support/commands/manual.py
into a focused amx/cli_support/_manual_errors.py.

The helpers turn raw DB driver exceptions into a short user-friendly
sentence + an actionable error line; the manual edit wizards pipe
every DB write failure through here.

manual.py re-exports both names so any caller importing the
underscore form (and catalog_picker.py's existing _ask_choice_or_cancel
imports nearby) keeps working unchanged.

manual.py: 1239 -> 1221 LOC (-18).
